Did some refactoring to clean up code and also added lots of comments
The loginToAccount function will now no longer return the access
token in the response, and the API spec will reflect this.

The API spec now refers to the Cookie Authentication security
scheme instead of the Bearer Authentication security scheme.

Deprecated the authenticateToken function.

Added another environment variable for the URL of the front-end web
application. This is a regular expression (without the slashes at
the start and end).

TODO: Remove the /testing route whenever possible