Created a route for checking if the user is logged in
It's a simple GET request and it simply returns the logged-in user's public info. It can be used for verifying if a user is logged in since it makes use of the authenticateTokenCookie function, which will return a non-successful status code if the user is not logged in.

Also changed some of the names of the jwt payload (user to username).

Blog posts will also now retrieve the username of the currently logged-in user.

API spec has been updated as needed.